requirements.txt
streamlit
pandas
requests
python-dotenv
plotly
psycopg2-binary
pytest
pytest-cov